Add RASPA init to conformer_search CLI.
RASPA simulation files can now be initialized using the CLI.
Files are copied from a given source directory and qsub script is
modified to change the job name. replace_lines.py is used to do that.

Default conformer output format is now 'mol' instead of 'pdb' since
there seems to an error with the pdb output of the MoleculeWriter.